requirements.txt
fastapi
pwdlib[argon2]
pydantic
pyjwt
python-multipart
uvicorn